HRI-63 - Moving the HMH constants to a events package and clean up
