+++ src/test/java/org/imsglobal/caliper/events/HMHConstants.java
@@ -1,10 +1,9 @@
-package org.imsglobal.caliper.ri.events;
+package org.imsglobal.caliper.events;
 
 public class HMHConstants {
 
     public static final String BASE_URN = "urn:uuid:";
     public static final String BASE_IRI = "https://www.hmhco.com/";
-    public static final String APP_NAME = "ReadingInventory";
     public static final String STUDENT_USER_REF_ID = "0f4fedbe-2227-415f-8553-40731a627171";
     public static final String ACTIVITY_REF_ID = "c050e852-5edb-4743-92d1-b53466de3a5f";
     public static final String SCHOOL_REF_ID = "ab570d90-3791-4048-a5ee-759657ddccef";
