+++ src/test/java/org/imsglobal/caliper/ri/events/GradeEventGradedTest.java
@@ -19,17 +19,18 @@
 package org.imsglobal.caliper.ri.events;
 
 import static com.yammer.dropwizard.testing.JsonHelpers.jsonFixture;
-import static org.imsglobal.caliper.ri.events.HMHConstants.ACTIVITY_REF_ID;
-import static org.imsglobal.caliper.ri.events.HMHConstants.APP_NAME;
-import static org.imsglobal.caliper.ri.events.HMHConstants.BASE_IRI;
-import static org.imsglobal.caliper.ri.events.HMHConstants.BASE_URN;
-import static org.imsglobal.caliper.ri.events.HMHConstants.DISTRICT_REF_ID;
-import static org.imsglobal.caliper.ri.events.HMHConstants.OBJECT_ID;
-import static org.imsglobal.caliper.ri.events.HMHConstants.SCHOOL_REF_ID;
-import static org.imsglobal.caliper.ri.events.HMHConstants.STUDENT_USER_REF_ID;
-import static org.imsglobal.caliper.ri.events.HMHConstants.LAST_ATTEMPT_ID;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+
+import static org.imsglobal.caliper.events.HMHConstants.ACTIVITY_REF_ID;
+import static org.imsglobal.caliper.events.HMHConstants.BASE_IRI;
+import static org.imsglobal.caliper.events.HMHConstants.BASE_URN;
+import static org.imsglobal.caliper.events.HMHConstants.DISTRICT_REF_ID;
+import static org.imsglobal.caliper.events.HMHConstants.LAST_ATTEMPT_ID;
+import static org.imsglobal.caliper.events.HMHConstants.OBJECT_ID;
+import static org.imsglobal.caliper.events.HMHConstants.SCHOOL_REF_ID;
+import static org.imsglobal.caliper.events.HMHConstants.STUDENT_USER_REF_ID;
+import static org.imsglobal.caliper.ri.events.RIConstants.APP_NAME;
 
 import org.imsglobal.caliper.TestUtils;
 import org.imsglobal.caliper.actions.Action;
