+++ src/test/java/org/imsglobal/caliper/ri/events/RIConstants.java
@@ -1,0 +1,7 @@
+package org.imsglobal.caliper.ri.events;
+
+public class RIConstants {
+	
+    public static final String APP_NAME = "ReadingInventory";
+
+}
